requirements.txt
Flask
flask-cors
pydicom
Pillow
numpy
